Split ErrorContainer into a seperate container and added default props and typechecking
